Resuelvo comentarios del PR; ordeno interfaces
