+++ src/components/exportable_card/FullCardLinks.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { ICardBaseConfig } from '../../indexCard';
-import { IWebSnippetOptions } from '../common/linkBuilders';
+import { IWebSnippetOptions } from './FullCardDropdown';
 import FullLinks from './links/FullLinks';
 import NullLinks from './links/NullLinks';
 import SmallLinks from './links/SmallLinks';
@@ -16,18 +16,13 @@ export interface ICardLinksOptions extends IWebSnippetOptions {
     downloadUrl: string
 }
 
-export default (props: {
-    serieId: string, cardOptions:ICardBaseConfig, downloadUrl: string
-}) => {
-    const LinkComponent = TYPES[props.cardOptions.links]
-    const Options = {
-        chartType: props.cardOptions.chartType,
-        color: props.cardOptions.color,
-        downloadUrl: props.downloadUrl,
-        hasChart: props.cardOptions.hasChart,
-        links: props.cardOptions.links,
-        serieId: props.serieId
-    }
+interface IFullCardLinksOptions extends ICardBaseConfig {
+    serieId: string,
+    downloadUrl: string
+}
 
-    return <div> <LinkComponent options={Options} /> </div>
+export default (props: { options: IFullCardLinksOptions }) => {
+    const LinkComponent = TYPES[props.options.links]
+
+    return <div> <LinkComponent options={props.options} /> </div>
 }
